+++ arduino/utils.h
@@ -1,6 +1,6 @@
 #ifndef _UTILS_H_
 #define _UTILS_H_
 
-double calculateAngularDistance(double targetAngle, double currentAngle, bool *clockwise);
+float calculateAngularDistance(float targetAngle, float currentAngle, bool *clockwise);
 
 #endif
